requirements.txt
requests
python-dotenv
urllib3
openai
azure-storage-blob
azure-search-documents
